requirements.txt
glcontext==2.5.0
moderngl==5.10.0
pygame-ce==2.4.1
